(refactor) code for get user profile - move user profile API call to controller.js - refactor the return data for user profile
