+++ server/db/user/index.js
@@ -7,18 +7,11 @@ var router = express.Router();
 var User = require('./user.model');
 var jwt = require('jsonwebtoken');
 var config = require('../../config/config');
+var controller = require('./controller');
 
 
 router
-  .get('/', function(req, res, next){
-    var decode = jwt.verify(JSON.parse(req.cookies.token), config.secrets.session);
-    User.findOne({
-      '_id': decode.id
-    },
-    function(err, user){
-      res.send(user);
-    });
-  })
+  .get('/', controller.getUserProfile)
   .get('/friends', function(req, res, next){
     var decode = jwt.verify(JSON.parse(req.cookies.token), config.secrets.session);
     User.findOne({
